refactor: SupplementalCIImagesConfig in separate file
